requirements.txt
anthropic
python-dotenv
requests
halo
